[api] Move the runtime log client out of the component package so the logs fast path stays light
