+++ script/ci-custom.py
@@ -557,7 +557,7 @@ def lint_constants_usage():
 # Maximum allowed CONF_ constants in esphome/const.py.
 # This file is frozen — new constants go in esphome/components/const/__init__.py.
 # Decrease this number when constants are moved out of const.py.
-CONST_PY_MAX_CONF = 1015
+CONST_PY_MAX_CONF = 1016
 
 
 @lint_content_check(include=["esphome/const.py"])
